@@ -2,21 +2,6 @@
   "name": "portofolio",
   "version": "0.1.0",
   "private": true,
-  "dependencies": {
-    "@material-ui/core": "^4.12.3",
-    "@material-ui/icons": "^4.11.2",
-    "@material-ui/styles": "^4.11.4",
-    "@testing-library/jest-dom": "^5.14.1",
-    "@testing-library/react": "^11.2.7",
-    "@testing-library/user-event": "^12.8.3",
-    "emailjs-com": "^3.2.0",
-    "framer-motion": "^4.1.17",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-router-dom": "^5.2.1",
-    "react-scripts": "4.0.3",
-    "web-vitals": "^1.1.2"
-  },
   "scripts": {
     "start": "react-dotenv && react-scripts start",
     "build": "react-dotenv && react-scripts build",
@@ -42,9 +27,25 @@
     ]
   },
   "react-dotenv": {
-    "whitelist": ["SERVICE_ID", "TEMPLATE_ID", "USER_ID"]
+    "whitelist": [
+      "SERVICE_ID",
+      "TEMPLATE_ID",
+      "USER_ID"
+    ]
   },
   "devDependencies": {
     "react-dotenv": "^0.1.3"
+  },
+  "dependencies": {
+    "@emotion/react": "^11.10.0",
+    "@emotion/styled": "^11.10.0",
+    "@mui/icons-material": "^5.8.4",
+    "@mui/material": "^5.10.1",
+    "emailjs": "^4.0.1",
+    "framer-motion": "^7.2.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.3.0",
+    "react-scripts": "^5.0.1"
   }
 }
